RoApi/requirements.txt
asgiref==3.7.2
distlib==0.3.6
Django==3.0
djangorestframework==3.14.0
filelock==3.12.2
importlib-metadata==6.7.0
Pillow==9.5.0
platformdirs==3.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
virtualenv==20.23.1
virtualenvwrapper-win==1.2.7
zipp==3.15.0
